fixed symlink script, additions to vimrc and git files

+++ scripts/shared/symlink.sh
@@ -19,16 +19,32 @@ toSymlinkDirectories=(
 )
 
 let n=0
-let total=${#toSymlinkFiles[@]}+${#toSymlinkDirectories[@]}
+let total=${#toSymlinkFiles[@]}+${#toSymlinkDirectories[@]}+1
+
+# Backups
+[ ! -d "$DOTFILES/backup" ] && mkdir -p $DOTFILES/backup
 
 for i in "${toSymlinkFiles[@]}"; do
 	progress_bar $n $total "Symlinking $i"
-	ln -sf $DOTFILES/link/$i ~/$i
+  if [[ -L ~/$i ]]; then
+    print_error "ERROR: ~/$i HAS ALREADY BEEN SYMBOLIC LINKED"
+  else
+    [ -f ~/$i ] && mv ~/$i $DOTFILES/backup/$i
+    ln -sf $DOTFILES/link/$i ~/$i
+  fi
 	((n++))
 done
 
 progress_bar $n $total "Symlinking .zsh"
-ln -sf $DOTFILES/link/.zsh ~/.zsh
+if [[ -L $HOME/.zsh ]]; then
+    print_error "ERROR: ~/.zsh/ HAS ALREADY BEEN SYMBOLIC LINKED"
+else
+  if [[ -d $HOME/.zsh ]]; then
+    mkdir -p $DOTFILES/backup/.zsh
+    mv $HOME/.zsh $DOTFILES/backup/.zsh
+  fi
+  ln -sf $DOTFILES/link/.zsh ~/.zsh
+fi
 ((n++))
 
 #Symbolic link ./.config
@@ -36,24 +52,49 @@ ln -sf $DOTFILES/link/.zsh ~/.zsh
 #Make directories
 for i in "${toSymlinkDirectories[@]}"; do
 	progress_bar $n $total "Symlinking $i"
-    path="*/*";
-    onlyOneWildcard=false;
-    if [[ "$i" = ".themes" ]] || [[ "$i" = ".icons" ]]
-    then
-        path="*";
-        onlyOneWildcard=true;
-    fi
-    for j in $DOTFILES/link/$i/$path; do
+    for j in $DOTFILES/link/$i/*/*; do
         relPath="$(echo $j | awk -F "/" '{ printf "%s/%s/%s", $6, $7, $8 }')"
         relPathBackOne="$(echo $j | awk -F "/" '{ printf "%s/%s", $6, $7}')"
-        echo $j
-        echo $HOME/$relPath
-        [ -d "$HOME/$relPath" ] && sudo rm -r $HOME/$relPath
-        [ ! -d "$HOME/$relPath" ] && mkdir -p $HOME/$relPathBackOne
-        sudo ln -sf $j $HOME/$relPath
+
+        print_info "Symlinking $HOME/$relPath"
+
+        if [[ -L $HOME/$relPath ]]; then
+          print_error "ERROR: $HOME/$relPath has already been symbolic linked"
+        else
+          # Backup and then symbolic link
+          [ ! -d "$HOME/$relPath" ] && mkdir -p $HOME/$relPathBackOne
+
+          if [[ -d "$HOME/$relPath" ]]; then
+            mkdir -p $DOTFILES/backup/$relPath
+            [ -d $HOME/$relPath ] && mv $HOME/$relPath $DOTFILES/backup/$relPathBackOne
+          elif [[ -f  "$HOME/$relPath" ]]; then
+            mkdir -p $DOTFILES/backup/$relPathBackOne
+            [ -f $HOME/$relPath ] && mv $HOME/$relPath $DOTFILES/backup/$relPath
+          fi
+
+          ln -sf $j $HOME/$relPath
+        fi
     done
-	((n++))
+
+    for j in $DOTFILES/link/$i/*; do
+        relPath="$(echo $j | awk -F "/" '{ printf "%s/%s/%s", $6, $7, $8 }')"
+        relPathBackOne="$(echo $j | awk -F "/" '{ printf "%s/%s", $6, $7}')"
+        # Backup and then symbolic link
+        if [[ -f $j ]]; then
+          print_info "Symlinking $HOME/$relPathBackOne"
+          if [[ -L $HOME/$relPathBackOne ]]; then
+            print_error "ERROR: $HOME/$relPath has already been symbolic linked"
+          else
+            if [[ -f  "$HOME/$relPath" ]]; then
+              mkdir -p $DOTFILES/backup/$relPathBackOne
+              [ -f $HOME/$relPath ] && mv $HOME/$relPath $DOTFILES/backup/$relPath
+            fi
+            ln -sf $j $HOME/$relPathBackOne
+          fi
+        fi
+    done
 done
+((n++))
 
 progress_bar $total $total "Done!"
 print_success "FINISHED SYMLINKING"
